Moved constants.py from tokenization/ to dataloaders/ for improved clarity.
