+++ data/mus/symbolic/tokenization/anticipation_tokenizer.py
@@ -3,7 +3,7 @@ import os
 from argparse import Namespace
 from pathlib import Path
 from path_utils import get_dataset_path, get_subset_path
-from tokenization.constants import DatasetType
+from dataloaders.constants import DatasetType
 
 # Import the main logic from your sub-modules
 from tokenization.anticipation.train.midi_preprocess import main as preprocess_main
